Bootstrap Python packaging
We add a setup.py with basic metadata and dependencies. Only "jinja2"
and "mitogen" dependencies are required. Other ones appear to be
optional: "import coloredlogs" is handled, and "parted" is only required
for the "device" module which is apparently not imported internally (the
dependency is thus defined as an "extra").

A MANIFEST.in file is also added to configure sdist. We exclude
developer scripts, probably not useful to distribute as is.